renamed vue component to avoid naming collision
