+++ src/components/index.ts
@@ -163,8 +163,8 @@ export { default as SchemaFormProperties } from './SchemaFormProperties.vue'
 export { default as SchemaFormProperty } from './SchemaFormProperty.vue'
 export { default as SchemaFormPropertyAllOf } from './SchemaFormPropertyAllOf.vue'
 export { default as SchemaFormPropertyAnyOf } from './SchemaFormPropertyAnyOf.vue'
-export { default as SchemaProperties } from './SchemaProperties.vue'
-export { default as SchemaProperty } from './SchemaProperty.vue'
+export { default as SchemaProperties } from './SchemaPropertiesKeyValues.vue'
+export { default as SchemaProperty } from './SchemaPropertyKeyValue.vue'
 export { default as SearchInput } from './SearchInput.vue'
 export { default as SelectedCount } from './SelectedCount.vue'
 export { default as SeparatedList } from './SeparatedList.vue'
