+++ src/types/index.ts
@@ -14,4 +14,4 @@ export * from './tabs'
 export * from './tailwind'
 export * from './UnionFilters'
 export * from './utilities'
-export type { SchemaValues } from './schemas'
+export * from './schemas'
